backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-dotenv
python-keycloak
